+++ c/vamk/vamk_i2c_switch_state_transmitter.h
@@ -10,7 +10,7 @@
 ///Function Declarations
 void i2c_switch_state_transmitter_init(void);
 void i2c_switch_state_request_handler(i2c_inst_t *i2c, i2c_slave_event_t event);
-void i2c_switch_state_on_switch_pressed(uint8_t row, uint8_t col);
-void i2c_switch_state_on_switch_released(uint8_t row, uint8_t col);
+void i2c_switch_state_on_switch_pressed(uint16_t row, uint16_t col);
+void i2c_switch_state_on_switch_released(uint16_t row, uint16_t col);
 
 #endif
